Organization of files
Files have been organized to correct the dealing of certain files. 

In the next version, flask navigation will be prepared so the website is no longer static. 

Then after that, work on databases and role based access would be worked on. 

Following that will be security features, and other features.